backend/requirements.txt
Flask==3.0.3
flask_cors==5.0.0
flask_pymongo==2.3.0
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
pytrends==4.9.2
Werkzeug==3.0.4
